test: split auth base rewrite tests
